Rename file img-text-column.vue to ImgTextColumn.vue
